Arrumou o botao de efeito sonoro Co-authored-by: anasofia0 <anasofialaranja@gmail.com>
