+++ Assets/Jogo/Interface/Opções.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6a5f2cb15550c089a2cd9ed9b74c1a3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
