+++ Assets/Scripts/DontDestroySound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 516ec301fc55dd072be8fb09d72cbc0f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
