+++ Assets/Scripts/MenuManager.cs
@@ -5,16 +5,6 @@ using UnityEngine;
 public class MenuManager : MonoBehaviour
 {
     public bool dialogue = true;
-    public GameObject musicaDeFundo;
-
-    void Awake()
-    {
-        if(PlayerPrefs.GetInt("Musica") == 1){
-            musicaDeFundo.SetActive(false);
-        } else {
-            musicaDeFundo.SetActive(true);
-        }
-    }
 
     public void BoolButtonDialogue()
     {
@@ -30,18 +20,5 @@ public class MenuManager : MonoBehaviour
             Debug.Log("dialogo true");
         }
     }
-
-    public void BoolButtonMusica()
-    {
-        if(musicaDeFundo.activeSelf){
-            musicaDeFundo.SetActive(false);
-            PlayerPrefs.SetInt("Musica", 1);
-            Debug.Log("Musica false");
-        } else {
-            musicaDeFundo.SetActive(true);
-            PlayerPrefs.SetInt("Musica", 0);
-            Debug.Log("Musica true");
-        }
-        PlayerPrefs.Save();
-    }
+    
 }
