+++ Assets/Scripts/MusicSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f12872be7e7124446be2bf6a6df60e70
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
