+++ Assets/Scripts/SoundEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 40678b417f4d6c04d9f59362dffd749a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
