+++ ProjectSettings/TagManager.asset
@@ -14,6 +14,7 @@ TagManager:
   - Ponte
   - Dialogo
   - Music
+  - Sons
   layers:
   - Default
   - TransparentFX
